waves 70-74: allocator consolidation, O(1) hash dispatch, tclhash unification
Wave 70: op_htab hash tables for bind table dispatch — find_bind_table()
and exact-match bind lookups are now O(1) instead of O(n).

Wave 71: op_htab for channel member lookup — ismember() is O(1).

Wave 72: allocator consolidation — all nmalloc/nrealloc/nfree call sites
(750+) migrated to op_malloc/op_realloc/op_free; mem.c removed; legacy
module Makefiles removed (meson-only build); egg_commio bridge added for
libop commio I/O layer; net.c reduced by ~900 lines.

Wave 73: op_htab for ban/exempt/invite mask lists — O(1) duplicate
detection and ismask() queries in channels.mod.

Wave 74: tclhash Tcl/no-Tcl dispatch unification — eliminated ~600 lines
of duplicated code via DISPATCH macro abstraction; fixed four latent bugs
in the no-Tcl path (check_tcl_chjn/chpt match type, check_tcl_dcc missing
BIND_QUIT, missing move-to-front optimisation).

118 files changed, ~3600 insertions, ~5580 deletions.

+++ libop/include/op_ratelimit.h
@@ -87,7 +87,7 @@ op_ratelimit_init(op_ratelimit_t *rl, unsigned capacity, unsigned rate_per_sec,
                   uint64_t now_usec)
 {
     rl->rate_usec   = (rate_per_sec > 0) ? 1000000u / rate_per_sec : UINT64_MAX;
-    rl->capacity    = capacity    * OP_RATELIMIT_SCALE;
+    rl->capacity    = (uint32_t)((uint64_t)capacity * OP_RATELIMIT_SCALE);
     rl->tokens      = rl->capacity;   /* start full */
     rl->last_refill = now_usec;
 }
@@ -149,11 +149,11 @@ op_ratelimit_check_n(op_ratelimit_t *rl, uint64_t now_usec, unsigned n)
         rl->tokens = (total > rl->capacity) ? rl->capacity : (uint32_t)total;
     }
 
-    uint32_t cost = n * OP_RATELIMIT_SCALE;
+    uint64_t cost = (uint64_t)n * OP_RATELIMIT_SCALE;
     if (rl->tokens < cost)
         return false;
 
-    rl->tokens -= cost;
+    rl->tokens -= (uint32_t)cost;
     return true;
 }
